mutations and querys

+++ books/models.py
@@ -5,7 +5,7 @@ class Book(models.Model):
     title = models.CharField(max_length=100)
     description = models.TextField()
     author = models.CharField(max_length=80)
-    price = models.DecimalField(max_digits=5, decimal_places=2)
+    price = models.FloatField()
     created_at = models.DateTimeField(auto_now_add=True)
     updated_at = models.DateTimeField(auto_now=True)
